fix: 엔티티 매핑 과정에서의 타입 불일치 해결 (#20)
- SMALLINT 타입 매칭

+++ src/main/java/com/raisedeveloper/server/domain/user/domain/UserCharacter.java
@@ -40,7 +40,7 @@ public class UserCharacter extends CreatedUpdatedEntity {
 	private String name;
 
 	@Column(nullable = false)
-	private int level;
+	private short level;
 
 	@Column(nullable = false)
 	private int exp;
